+++ src/ua/data_types/browse_request.rs
@@ -1,4 +1,4 @@
-use crate::{ua, ServiceRequest};
+use crate::{ua, DataType as _, ServiceRequest};
 
 crate::data_type!(BrowseRequest);
 
@@ -22,4 +22,12 @@ impl BrowseRequest {
 
 impl ServiceRequest for BrowseRequest {
     type Response = ua::BrowseResponse;
+
+    fn request_header(&self) -> &ua::RequestHeader {
+        ua::RequestHeader::raw_ref(&self.0.requestHeader)
+    }
+
+    fn request_header_mut(&mut self) -> &mut ua::RequestHeader {
+        ua::RequestHeader::raw_mut(&mut self.0.requestHeader)
+    }
 }