+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Display and parse `ua::ExpandedNodeId`.
+- Perform generic, asynchronous read requests with custom response timeout.
 
 ## [0.9.2] - 2025-07-15